Move the var-map struct in a separate file.
